requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.10
Django==5.1.3
django-crispy-forms==2.3
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
openai>=1.30.0
google-generativeai>=0.8.0
